Alter store structure
