+++ frontend/package.json
@@ -29,7 +29,7 @@
   },
   "devDependencies": {
     "@vue/cli": "^3.7.0",
-    "@vue/cli-plugin-babel": "^3.7.0",
+    "@vue/cli-plugin-babel": "^3.8.0",
     "@vue/cli-plugin-eslint": "^3.7.0",
     "@vue/cli-plugin-unit-jest": "^3.7.0",
     "@vue/cli-service": "^3.7.0",
